moved fragments into the components folder - renamed Titleable - added in better at to queue images - worked more on getting the titling working correctly - it all works now except for when you press back from the menu
