+++ fanClub/res/layout/song_item.xml
@@ -45,13 +45,14 @@
         android:textSize="14sp" >
     </TextView>
     
-    <Button 
+    <ImageButton
         android:id="@+id/add_to_playlist"
+        android:src="@drawable/av_add_to_queue"
         android:layout_width = "wrap_content"
         android:layout_height="wrap_content"
         android:gravity="center_vertical"
         android:layout_alignParentRight="true"
         android:visibility="invisible"
-        android:background="@drawable/add_to_queue"/>
+        android:background="@null"/>
 
 </RelativeLayout>
