+++ fanClub/src/com/lastcrusade/fanclub/util/ITitleable.java
@@ -1,6 +1,6 @@
 package com.lastcrusade.fanclub.util;
 
-public interface Titleable {
+public interface ITitleable {
     
     public abstract String getTitle();
